+++ src/stdlib_utils/queue_utils.py
@@ -11,24 +11,21 @@ import multiprocessing.queues
 import queue
 from queue import Empty
 from queue import Queue
+import time
 from time import process_time
 from typing import Any
 from typing import List
 from typing import Union
 
+from .constants import SECONDS_TO_SLEEP_BETWEEN_CHECKING_QUEUE_SIZE
+from .constants import UnionOfThreadingAndMultiprocessingQueue
+from .exceptions import QueueNotExpectedSizeError
 from .exceptions import QueueStillEmptyError
 
 
 def _eventually_empty(
     should_be_empty: bool,
-    the_queue: Union[
-        Queue[  # pylint: disable=unsubscriptable-object # Eli (3/12/20) not sure why pylint doesn't recognize this type annotation
-            Any
-        ],
-        multiprocessing.queues.Queue[  # pylint: disable=unsubscriptable-object # Eli (3/12/20) not sure why pylint doesn't recognize this type annotation
-            Any
-        ],
-    ],
+    the_queue: UnionOfThreadingAndMultiprocessingQueue,
     timeout_seconds: Union[float, int] = 0.2,
 ) -> bool:
     """Help to determine if queue is eventually empty or not."""
@@ -40,18 +37,16 @@ def _eventually_empty(
             value_to_check = not value_to_check
         if value_to_check:
             return True
+        time.sleep(
+            SECONDS_TO_SLEEP_BETWEEN_CHECKING_QUEUE_SIZE
+        )  # wait 50 msec between polling the queue to allow things to process in the background
+        # process_time() does not include time during sleep, so decrement the timeout_seconds to account for this
+        timeout_seconds -= SECONDS_TO_SLEEP_BETWEEN_CHECKING_QUEUE_SIZE
     return False
 
 
 def is_queue_eventually_empty(
-    the_queue: Union[
-        Queue[  # pylint: disable=unsubscriptable-object # Eli (3/12/20) not sure why pylint doesn't recognize this type annotation
-            Any
-        ],
-        multiprocessing.queues.Queue[  # pylint: disable=unsubscriptable-object # Eli (3/12/20) not sure why pylint doesn't recognize this type annotation
-            Any
-        ],
-    ],
+    the_queue: UnionOfThreadingAndMultiprocessingQueue,
     timeout_seconds: Union[float, int] = 0.2,
 ) -> bool:
     """Check if queue is empty prior to timeout occurring."""
@@ -59,14 +54,7 @@ def is_queue_eventually_empty(
 
 
 def is_queue_eventually_not_empty(
-    the_queue: Union[
-        Queue[  # pylint: disable=unsubscriptable-object # Eli (3/12/20) not sure why pylint doesn't recognize this type annotation
-            Any
-        ],
-        multiprocessing.queues.Queue[  # pylint: disable=unsubscriptable-object # Eli (3/12/20) not sure why pylint doesn't recognize this type annotation
-            Any
-        ],
-    ],
+    the_queue: UnionOfThreadingAndMultiprocessingQueue,
     timeout_seconds: Union[float, int] = 0.2,
 ) -> bool:
     """Check if queue is not empty prior to timeout occurring."""
@@ -74,14 +62,7 @@ def is_queue_eventually_not_empty(
 
 
 def is_queue_eventually_of_size(
-    the_queue: Union[
-        Queue[  # pylint: disable=unsubscriptable-object # Eli (3/12/20) not sure why pylint doesn't recognize this type annotation
-            Any
-        ],
-        multiprocessing.queues.Queue[  # pylint: disable=unsubscriptable-object # Eli (3/12/20) not sure why pylint doesn't recognize this type annotation
-            Any
-        ],
-    ],
+    the_queue: UnionOfThreadingAndMultiprocessingQueue,
     size: int,
     timeout_seconds: Union[float, int] = 0.2,
 ) -> bool:
@@ -95,19 +76,33 @@ def is_queue_eventually_of_size(
     while process_time() - start_time < timeout_seconds:
         if the_queue.qsize() == size:
             return True
+        time.sleep(
+            SECONDS_TO_SLEEP_BETWEEN_CHECKING_QUEUE_SIZE
+        )  # wait 50 msec between polling the queue to allow things to process in the background
+        # process_time() does not include time during sleep, so decrement the timeout_seconds to account for this
+        timeout_seconds -= SECONDS_TO_SLEEP_BETWEEN_CHECKING_QUEUE_SIZE
     return False
+
+
+def confirm_queue_is_eventually_of_size(
+    the_queue: UnionOfThreadingAndMultiprocessingQueue,
+    size: int,
+    timeout_seconds: Union[float, int] = 0.2,
+) -> None:
+    """Raise exception if queue is not a certain size prior to timeout.
+
+    Typically used in unit testing to ensure that a put or get operation
+    has fully completed during test setup before triggering the function
+    being tested.
+    """
+    if is_queue_eventually_of_size(the_queue, size, timeout_seconds=timeout_seconds):
+        return
+    raise QueueNotExpectedSizeError(the_queue, size)
 
 
 def put_object_into_queue_and_raise_error_if_eventually_still_empty(  # pylint: disable=invalid-name # Eli (10/22/20): I know this is long, but it's a combined helper function for unit testing
     obj: object,
-    the_queue: Union[
-        Queue[  # pylint: disable=unsubscriptable-object # Eli (3/12/20) not sure why pylint doesn't recognize this type annotation
-            Any
-        ],
-        multiprocessing.queues.Queue[  # pylint: disable=unsubscriptable-object # Eli (3/12/20) not sure why pylint doesn't recognize this type annotation
-            Any
-        ],
-    ],
+    the_queue: UnionOfThreadingAndMultiprocessingQueue,
     timeout_seconds: Union[float, int] = 0.2,
 ) -> None:
     """Put an object into a queue and wait until queue is populated.